Create export_data_csvs
Hi team,

This script pulls all of the inspection data and corresponding yelp review info if available by phone number. I have successfully tested it and looped through the first 500 phone numbers today. We can tie this together with the script to add these csvs to the storage, and simply change the range of the loop each time to obtain a new batch of reviews. Thanks for your help an patience!